test: skip live-LLM E2E tests gracefully when no credentials are set
The keyless community CI lane (and any local run without secrets) hit a panic
or a live-API error instead of a skip: several E2E/integration tests require
live LLM and/or OpenAI-embedding credentials but did not detect their absence.
This violated the repo's documented "tests gracefully skip when env vars
unavailable" convention. Only the keyed lane (which always had a key) ever
exercised them, so the gap stayed latent until the secret-free lane was added.

Two failure modes are fixed:

1. require_env panic — tests that call `require_env("OPENAI_*")` (directly or via
   the adapter builder) at the top panic when the var is unset. Fixed by an
   early `llm_env_available()` guard + skip notice.

2. Lazy-OpenAI embedding 401 — after main switched tests to OpenAI embeddings
   by default, `create_test_embedding_engine()` constructs the OpenAI engine
   lazily (no key validation), so tests relying on it to skip instead failed
   with HTTP 401 at embed time (e2e_memify, integration_embeddings). The same
   credential guard turns that into a clean skip.

Shared helpers (dedup): the copy-pasted `require_env`, `llm_env_available`, the
model-default expression, and `create_adapter_from_env` now live once in
`cognee-test-utils` (`require_env`, `llm_env_available`, `llm_model_from_env`,
`create_openai_adapter_from_env`). Per-crate `tests/test_utils.rs` re-export them
under the historical names so existing call sites are untouched. This also
removes a review-flagged inconsistency: the model is now always defaulted (never
`require_env`'d) and every `require_env` honors the `LLM_*` aliases, so
`llm_env_available()` passing always implies the adapter builds — no more panic
on `LLM_*`-only creds or an absent OPENAI_MODEL. Net -224 lines across the tests.

Scope note: cassette-converted tests already run offline under
COGNEE_TEST_REPLAY=1 with no credentials, so they need no guard and are
untouched.
