+++ crates/cognify/tests/e2e_ontology_pipeline.rs
@@ -15,7 +15,7 @@ use cognee_database::{DatabaseConnection, IngestDb, SearchHistoryDb, connect, in
 use cognee_embedding::{EmbeddingEngine, MockEmbeddingEngine};
 use cognee_graph::{GraphDBTrait, LadybugAdapter};
 use cognee_ingestion::AddPipeline;
-use cognee_llm::{Llm, OpenAIAdapter};
+use cognee_llm::Llm;
 use cognee_models::DataInput;
 use cognee_ontology::{OntologyFileInput, OntologyResolver, RdfLibOntologyResolver};
 use cognee_search::{
@@ -30,7 +30,6 @@ use tempfile::TempDir;
 use uuid::Uuid;
 
 mod test_utils;
-use test_utils::require_env;
 
 const ONTOLOGY_FIXTURE: &str = "tests/test_data/ontology/tech_taxonomy.ttl";
 const ONTOLOGY_TECH_ONLY_FIXTURE: &str = "tests/test_data/ontology/tech_only.ttl";
@@ -118,9 +117,10 @@ fn search_contains_any(response: &SearchResponse, markers: &[&str]) -> bool {
 
 #[tokio::test]
 async fn e2e_ontology_pipeline_add_cognify_search() {
-    let _ = require_env("OPENAI_URL");
-    let _ = require_env("OPENAI_TOKEN");
-    let _ = require_env("OPENAI_MODEL");
+    if !test_utils::llm_env_available() {
+        eprintln!("skipping: live LLM credentials (OPENAI_URL/OPENAI_TOKEN) not set");
+        return;
+    }
 
     let temp_dir = TempDir::new().expect("temp dir");
 
@@ -153,14 +153,9 @@ async fn e2e_ontology_pipeline_add_cognify_search() {
     // In-memory mock vector DB (qdrant extracted to closed cognee-vector-qdrant).
     let vector_db: Arc<dyn VectorDB> = Arc::new(MockVectorDB::new());
 
-    let llm: Arc<dyn Llm> = Arc::new(
-        OpenAIAdapter::new(
-            require_env("OPENAI_MODEL"),
-            require_env("OPENAI_TOKEN"),
-            Some(require_env("OPENAI_URL")),
-        )
-        .expect("OpenAIAdapter::new"),
-    );
+    // Real OpenAI-compatible adapter (endpoint/key/model from env). Guarded
+    // above by `llm_env_available()`, so this never panics on the keyless lane.
+    let llm: Arc<dyn Llm> = test_utils::create_adapter_from_env();
 
     let owner_id = Uuid::nil();
